formatting loops in cheetah template with spacing
The cheetah template flow control is used to format the document for
assignments/solutions with variable number of problems. Preliminary
support for adding vertical spacing between problems in assignments is
given, however, it remains to ensure that there is a minimum vertical
spacing only in the body section of the document. As implemented the
spacing considers bottom and top margins which gives too little space
for problems near the bottom of pages