in-place methods and passing class to string fmt
all methods are now in-place to give consistency to the classes in the
module. an advantage of this is that the variable class can be passed to
the string format, so that more than just its value can be used. this is
an advantage if one wants to access any of the existing attributes, such
as lower bound and upper bound, or possibly for future development
units, aliases (to randomize the symbols used to represent a quantity,
especially important for questions asking for derivations or symbolic
expressions as the answer)