WIP support for quantities, randomize units
current implementation is inelegant/inefficient, defining wrapper class
with awkward methods since the quantities package doesn't have an API to
all of its modules. this can be fixed by understanding the package and
finding the correct objects to import and use, and refactoring the
current classes (a Units class may not be necessary, though it with its
own in-place rng method is consistent with the existing class design)

randomized units introduce the problem of conversion factors introducing
many non-zero values down to machine precision. while it is possible to
define the rounding in string formatting methods like {x.value:.3f},
unless using exponential notation that is not precision but decimal
place specification.

this commit is saved for the following reason: the problem is to
keep consistent with the integer class which isn't expected to have
units, the value attribute even when assigned a quantity should have a
string method returning only the magnitude. this is also important for
formatting in latex since units are generally in rm font, not italics.
but redefining the string method for Float in problem.py, while succesful in
the module execution (see if __name__ == '__main__' conditional), is
unsuccesful in output generated by compile.py