add container classes
container classes are made, and the convention for outputs is now to use
the container classes, designed for inputs and units. this makes all
quantities share the same classes and methods to avoid bugs

to be more brief and omit .value in the text formatting for variables and their units,
the string methods for the container classes is made to just print the value
(magnitude, in the case of a variable, and the unit string, in the case
of a variable.unit).

quantities are not explicitly supported, instead just unit quantities
being used as conversion factors. if desired it is possible to construct
quantitiy objects from the container classes used as inputs to the
solver and then use some of its convenience functions such as simplify

rounding for precision is also introduced