support symbolic randomization
symbol randomization for inputs is now supported. previously the solver
was just being passed the value attribute, now it is being passed the
whole object. this is necessary for, e.g., the units on float variables
to be accessed in the solver function body

likely refactoring of the current objects is required to define
multiplication. while it was possible to eliminate the .value
specification in the string formats by defining a __str__ method,
similarly all multiplication and other methods must be defined for the
objects if they are to be a superclass of a numpy array, which seems to
be most appropriate