first support for randomized array size
this is implemented where the user doesn't have to explicilty call
the object, it is inferred from the input data structure to the size
call. that design choice is inconsistent with previous, but it could be
applied to all, for example, a dimensionless random float variable in an
interval is one for which the value is specified as a 3-tuple, all of
which are floats. a constant float with units is a 2-tuple with float at
first index, string in second

also there is high redundancy because everything that is randomized
has to be put into a container with a randomization method, even in
cases where there is only one value so randomization does nothing. the
alternative is to check if there is a randomization method and only call
it if it exists