dimension evaluation changed to unit parse
dictionary data type is used to parse any units including
mixed units such as "ft*m/(hr*min)"

only power supports integer and integer operation. currently constant
scale factors are not supported, for example "thousands of tons" or
1000*ton, since they would require the linear addition and subtraction
operations to keep multiplicative constants. however this can be
sidestepped by defining the unit "1000*tons" in the dictionaries,
effectively using prefixes like the greek metric prefixes

the units dictionary is very similar to the dimensionality
attribute used in python-quantities. the units module of the brian2
package actually appears to be fuller featured, and defines operations
on the dimension object as 7-tuple analogous to those used here in
interpreting the string:

    def __mul__(self, value):
        return get_or_create_dimension([x + y for x, y in
                                        zip(self._dims, value._dims)])

    def __div__(self, value):
        return get_or_create_dimension([x - y for x, y in
                                        zip(self._dims, value._dims)])

    def __truediv__(self, value):
        return self.__div__(value)

    def __pow__(self, value):
        value = np.array(value, copy=False)
        if value.size > 1:
            raise TypeError('Too many exponents')
        return get_or_create_dimension([x * value for x in self._dims])