first implementation of a Quantity object
The view method is used on the numpy array to create a subclass of the
ndarray. as the readme says, this quantities object is a simplified
version of the brian2 quantities object, in which there are no other
non-builtin classes other than numpy arrays (the units are a dictionary
and the dimension is a numpy array).

the tests are not exhaustive and the methods which make this a suitable
object, such as a randomization, remain to be implemented